+++ src/client/index.js
@@ -1,12 +1,5 @@
 import { render } from './render';
-import errorOverlay from 'vscode-notebook-error-overlay';
-import type { ActivationFunction } from 'vscode-notebook-renderer';
-
-// Fix the public path so that any async import()'s work as expected.
-declare const __webpack_relative_entrypoint_to_root__: string;
-declare const scriptUrl: string;
-
-__webpack_public_path__ = new URL(scriptUrl.replace(/[^/]+$/, '') + __webpack_relative_entrypoint_to_root__).toString();
+import { ActivationFunction } from 'vscode-notebook-renderer';
 
 // ----------------------------------------------------------------------------
 // This is the entrypoint to the notebook renderer's webview client-side code.
@@ -15,7 +8,8 @@ __webpack_public_path__ = new URL(scriptUrl.replace(/[^/]+$/, '') + __webpack_re
 // rendering logic inside of the `render()` function.
 // ----------------------------------------------------------------------------
 
-export const activate: ActivationFunction = context => {
+/** @type {ActivationFunction} */
+export function activate(context) {
   return {
     renderOutputItem(outputItem, element) {
       let shadow = element.shadowRoot;
@@ -25,15 +19,14 @@ export const activate: ActivationFunction = context => {
         root.id = 'root';
         shadow.append(root);
       }
-      const root = shadow.querySelector<HTMLElement>('#root')!;
-      errorOverlay.wrap(root, () => {
-        root.innerHTML = '';
-        const node = document.createElement('div');
-        root.appendChild(node);
+      const root = shadow.querySelector('#root');
+      root.innerHTML = '';
+      const node = document.createElement('div');
+      root.appendChild(node);
 
-        render({ container: node, mime: outputItem.mime, value: outputItem.json(), context });
-      });
+      render({ container: node, mime: outputItem.mime, value: outputItem.json(), context });
     },
+    
     disposeOutputItem(outputId) {
       // Do any teardown here. outputId is the cell output being deleted, or
       // undefined if we're clearing all outputs.